refactors for comments
