+++ src/cryptography/hazmat/bindings/_rust/openssl/hashes.pyi
@@ -17,3 +17,11 @@ class Hash(hashes.HashContext):
     def copy(self) -> Hash: ...
 
 def hash_supported(algorithm: hashes.HashAlgorithm) -> bool: ...
+
+class XOFHash:
+    def __init__(self, algorithm: hashes.ExtendableOutputFunction) -> None: ...
+    @property
+    def algorithm(self) -> hashes.ExtendableOutputFunction: ...
+    def update(self, data: bytes) -> None: ...
+    def squeeze(self, length: int) -> bytes: ...
+    def copy(self) -> XOFHash: ...
