--- src/cryptography/hazmat/bindings/_rust/openssl/xofhash.pyi
@@ -1,15 +1,0 @@
-# This file is dual licensed under the terms of the Apache License, Version
-# 2.0, and the BSD License. See the LICENSE file in the root of this repository
-# for complete details.
-
-from cryptography.hazmat.primitives import hashes
-
-class XOFHash:
-    def __init__(self, algorithm: hashes.ExtendableOutputFunction) -> None: ...
-    @property
-    def algorithm(self) -> hashes.ExtendableOutputFunction: ...
-    def update(self, data: bytes) -> None: ...
-    def squeeze(self, length: int) -> bytes: ...
-    def copy(self) -> XOFHash: ...
-
-def xofhash_supported(algorithm: hashes.ExtendableOutputFunction) -> bool: ...
